Implement process_3d
