+++ lang/model/operation.rs
@@ -33,27 +33,6 @@ pub trait Operation: std::fmt::Debug {
     }
 }
 
-impl Operation for BooleanOp {
-    fn process_2d(&self, model: &Model) -> Geometries2D {
-        let mut geometries = Geometries2D::default();
-
-        if let Some(model) = model.into_inner_object_model() {
-            let self_ = model.borrow();
-            self_.children.iter().for_each(|model| {
-                let b = model.borrow();
-                let mat = b.output.local_matrix_2d();
-                geometries.append(
-                    model
-                        .process_2d(model)
-                        .transformed_2d(&self_.output.resolution, &mat),
-                );
-            });
-        }
-
-        geometries.boolean_op(&model.borrow().output.resolution, self)
-    }
-}
-
 /// Transformation matrix
 #[derive(Clone, Debug)]
 pub enum AffineTransform {
